*: break out functionality into multiple files
this provides better structure for the functionality so that I'm not
just working in one giant main file the whole time.
